backend/requirements.txt
pywebpush==2.0.0
python-jose[cryptography]==3.3.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
redis==5.0.4

# transformers v4.x series for stability
transformers==4.40.1
sentence-transformers==3.0.0

--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
pillow==10.3.0

python-dotenv==1.0.1
pandas==1.5.3
pydantic[email]==2.7.1

numpy==1.26.4
scikit-learn==1.4.0
slowapi==0.1.9
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
requests==2.31.0
mtranslate==1.8
waitress==3.0.0
python-multipart==0.0.9
authlib==1.3.0
sentry-sdk[fastapi]==2.1.1
itsdangerous==2.1.2
groq==0.5.0
moviepy==1.0.3
resend==0.8.0
razorpay==1.4.1
tenacity==8.3.0
alembic>=1.13.1
boto3==1.34.106
python-json-logger>=3.0.0
diffusers==0.27.2
accelerate==0.30.1
TTS==0.22.0
openai-whisper==20231117
pydub==0.25.1
firebase-admin==6.5.0
google-cloud-firestore==2.15.0
celery[redis]==5.3.6
flower==2.0.0
prometheus-fastapi-instrumentator==7.0.0
together==1.1.0
faiss-cpu==1.8.0
llama-cpp-python==0.2.76
PyPDF2==3.0.1
python-docx==1.1.2
tavily-python==0.3.3
google-cloud-run==0.10.5
google-cloud-tasks==2.16.4
google-cloud-storage==2.16.0
gcsfs==2024.3.1
neo4j==5.20.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
tiktoken==0.7.0
passlib[bcrypt]==1.7.4
openai==1.30.1

# Hardware & Device Integration
PyAudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
librosa==0.10.1
nvidia-ml-py==12.535.133
psutil==5.9.8
aiohttp==3.9.5
